feat: change package path of autoconfigure
