+++ object-storage-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -15,6 +15,6 @@
 # limitations under the License.
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.github.anilople.object.storage.config.ObjectStorageAutoConfiguration,\
-com.github.anilople.object.storage.config.S3ClientAutoConfiguration,\
-com.github.anilople.object.storage.config.S3PresignerAutoConfiguration
+com.github.anilople.object.storage.autoconfigure.ObjectStorageAutoConfiguration,\
+com.github.anilople.object.storage.autoconfigure.S3ClientAutoConfiguration,\
+com.github.anilople.object.storage.autoconfigure.S3PresignerAutoConfiguration
